+++ NEWS.md
@@ -1,3 +1,26 @@
+# ARDL 0.2.3
+
+### New features
+
+* Add new function `to_lm()` which converts `ardl`, `uecm`, and `recm` models, 
+as well as `dynlm` models in `lm` models. Among other things, this helps using 
+the model with other functions (e.g. with `predict()` to forecast)
+(issue #5, @melville1808).
+
+* `multipliers()` now has a new argument (`se`) which allows the estimation of
+standard errors for the delay multipliers, using the delta method.
+(issue #17, @sammy-w)
+
+* Add new function `plot_delay()` which creates plots (`ggplot`) for the delay 
+multipliers and their uncertainty intervals (standard errors or confidence 
+intervals).
+
+* Add new function `plot_lr()` which creates a plot (`ggplot`) for the long-run 
+relationship in comparison with the dependent variable, and the fitted values of
+the model.
+
+---
+
 # ARDL 0.2.2
 
 ### New features